test-1/requirements.txt
python-telegram-bot
g4f
